feat: add simple runtime, future and task implementation
Simple runtime is a FIFO polling queue which it's performend in a single
thread. All of these operations are backed up by ucontext_t which it allows
switch contexts, save program counter and stack where given task was yielded.

There are 2 common async functions defined in common.h, those are sleep and
join, but it will be moved in other place once the whole implementation and
API are well defined.