feat(runtime): add context which runtime is running
When a runtime starts its event loop, it will push itself on
the runtime context stack, which functions as taskio_runtime_spawn
and taskio_runtime_spawn_blocking will be allowed to use the current
runtime.

If there's no runtime is said stack, it will occur a segmentation fault
due to using a NULL pointer.