feat(runtime): implement a simple waker for simple runtime
Waker is an important piece for the runtime itself, which it allows
to poll when it be needed to do.

Waker works in the following way:

  - Runtime chooses who polls, creates the waker for that poll.

  - If future is pending still, it won't drop the waker because
    future must store it somewhere to be able to resume the future
    and runtime.

  - If future has finished, then no wake is needed and dropping
    is now needed to do due to future won't store the waker
    when it's ready.

Before when there were no ready futures to poll, the runtime would
shutdown. But that changed because now it can have sleeping futures
waiting to be woken and no ready futures to poll.