feat: create taskio_defer*(...) to drop memory when aborting task
When allocating memory in heap inside an async function, it should
be freed when given future is cancelled or aborted, to do that so,
users can now push on future itself the function which frees that data
avoiding memory leaks.

The taskio_drop_node struct works as a stack, freeing memory from
bottom to top.